pyproject.toml
[project]
name = "lukaton"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-adk>=1.27.1",
    "pdfplumber>=0.11.9",
    "pikepdf>=9.0",
    "python-dotenv>=1.0",
]
